refactor: Revise Child Components to BottomCenter,Right and Center, TopLeft template component
The role of template component is to layout or place the modules, organisms. But the problem was every page has slightly different template. So I decided to create single template component rather than focusing on reusability.
The ChildBig component is now CenterTemplate, ChildSmall is now BottomRight, ChildCenter is now BottomCenter, ChildTopLeft is now TopLeft template component.